Fixes #328
